chore(rafty): Remove old storage and logs funcs
Related to gihub issue #61
