+++ logs_types.go
@@ -1,9 +1,5 @@
 package rafty
 
-import (
-	"github.com/Lord-Y/rafty/raftypb"
-)
-
 // LogKind represent the kind of the log
 type LogKind uint8
 
@@ -30,58 +26,6 @@ const (
 	// No stale read can happen in this mode.
 	LogCommandLinearizableRead
 )
-
-// logs holds all requirements to manipulate logs
-type logs struct {
-	// rafty holds rafty config
-	rafty *Rafty
-
-	// log holds all logs entries
-	log []*raftypb.LogEntry
-}
-
-// logOperationReadResponse will be used by logOperationRequest.responseChan
-type logOperationReadResponse struct {
-	// Kind is the log operation kind
-	logs []*raftypb.LogEntry
-
-	// total is the total current total of logs
-	total int
-
-	// err return an error if there is one
-	err error
-}
-
-// logOperationReadResponse will be used by logOperationRequest.responseChan
-type logOperationReadLastLogResponse struct {
-	// Kind is the log operation kind
-	logs []*raftypb.LogEntry
-
-	// total is the total current total of logs
-	total int
-
-	// sendSnapshot tell us if it's better to send a snapshot
-	// instead of catchup logs
-	sendSnapshot bool
-
-	// lastLogIndex is the last log index to use when sending the catchup entries
-	lastLogIndex uint64
-
-	// lastLogTerm is the last log term of the logs from lastLogIndex
-	lastLogTerm uint64
-
-	// err return an error if there is one
-	err error
-}
-
-// logOperationWipeResponse will be used by logOperationRequest.responseChan
-type logOperationWipeResponse struct {
-	// total is the total current total of logs
-	total int
-
-	// err return an error if there is one
-	err error
-}
 
 // LogEntry is holds requirements that will be used
 // to store logs on disk
